feat: Header 파일 분리 및 상태 호이스팅 적용
